// src/app/shared/interfaces.ts
export type TsActivityType = 'running' | 'cycling' | 'bikepacking' | 'fastpacking' | 'skitouring' | 'canicross' | 'walking' | 'gravelbiking';
// ...
export interface TsListItem {
  name: string;
  stats: TsPathStats;
  info: TsInfo;
  pathId: string;
  isActive?: boolean;
  colour?: string;
  creationDate?: string;
  activityType?: string;
  isPublic?: boolean;
  // startPoint?: TsPosition;
}
// ...
export type TsPosition = [number, number];
// ...
export interface TsInfo {
    direction: string;
    category: string;
    nationalTrail: boolean;
    name: string;
    description: string;
    pathType: string;           // 'route' or 'track'
    startTime: string;
    isLong: boolean;
    isPublic: boolean;
    createdBy: string;
    isElevations: boolean;
    activityType: TsActivityType;
    startPoint: TsPosition;
}

export interface TsParams {
    elevs: number[];
    time: number[];
    heartRate: number[];
    cadence: number[];
    cumDistance: number[];
    matchedPoints?: number[][]; /// used for debugging route algorthims on the back end
}

export interface TsPathStats {
    bbox: {
        minLng: number,
        minLat: number,
        maxLng: number,
        maxLat: number
    };
    nPoints: number;
    simplificationRatio: number;
    duration: number;
    distance: number;
    pace:  number;
    elevations: {
        ascent: number,
        descent: number,
        maxElev: number,
        minElev: number,
        lumpiness: number,
        distance: number,
        nPoints: number,
        badElevData?: boolean
    };
    p2p: {
        max: number,
        ave: number
    };
    movingStats: {
        movingTime: number,
        movingDist: number,
        movingPace: number,
    };
    hills: TsHill[];
    splits: {
        kmSplits: number[][],
        mileSplits: number[][]
    };
}


// interface TsHills {

//   dHeight?: number;
//   dDist?: number;
//   dTime?: number;
//   pace?: number;
//   ascRate?: number;
//   gradient?: {
//       max?: number;
//       ave?: number
//   };
// }

interface TsHill {
  dHeight: number;
  dDist: number;
  startDist: number;
  startPoint: number;
  endPoint: number;
  dTime: number;
  pace: number;
  ascRate: number;
  maxGrad: number;
  aveGrad: number;
}